Adicionados componentes OrderTable e OrderForms | Implementados data-testids da pagina Checkout | Falta implementar lógica que atualiza o preco total
