+++ front-end/src/Components/Organisms/LoginForm.js
@@ -40,13 +40,21 @@ function LoginFormComponent({ title }) {
     setLoginForm({ ...loginForm, [name]: value });
   };
 
-  //  Teste
   const logIn = async () => {
     const response = await userLogin({ email: login, password });
-    return response.message
-      ? setErrorMessage({
+    if (response.message) {
+      return setErrorMessage({
         message: 'Login ou senha inválidos! :(',
-      }) : history.push(`/${response.role}/products`);
+      });
+    }
+    localStorage.setItem('user', JSON.stringify(response));
+    if (response.role === 'customer') {
+      history.push('/customer/products');
+    } else if (response.role === 'seller') {
+      history.push('/seller/orders');
+    } else {
+      history.push('/admin');
+    }
   };
 
   return (
